+++ packages/openapi-bundler/src/bundle.ts
@@ -3,20 +3,16 @@ import { oas30 } from "openapi3-ts";
 import { _, File, Folder } from "@dasaplan/ts-sdk";
 import { appLog } from "./logger.js";
 import { createSpecProcessor } from "./post-process/index.js";
+import { PostProcessingOptions } from "./post-process/post-process.js";
 
 export interface OpenApiBundled extends oas30.OpenAPIObject {}
 
-interface BundleOption {
+interface BundleOption extends PostProcessingOptions {
   postProcessor: (bundled: OpenApiBundled) => OpenApiBundled;
   outFile: string;
-  ensureDiscriminatorValues: boolean;
-  mergeAllOf: boolean;
 }
 
-export async function bundleOpenapi(
-  _pathToApi: string,
-  params?: Partial<BundleOption>
-) {
+export async function bundleOpenapi(_pathToApi: string, params?: Partial<BundleOption>) {
   appLog.childLog(bundleOpenapi).info("start bundle: ", _pathToApi);
   const inputFile = File.of(_pathToApi);
 
@@ -29,10 +25,7 @@ export async function bundleOpenapi(
   return { parsed, outFile: outputFile.writeYml(parsed) };
 }
 
-export async function bundleParseOpenapi(
-  _pathToApi: string,
-  params?: Partial<Omit<BundleOption, "outFile">>
-): Promise<OpenApiBundled> {
+export async function bundleParseOpenapi(_pathToApi: string, params?: Partial<Omit<BundleOption, "outFile">>): Promise<OpenApiBundled> {
   const inputFile = File.of(_pathToApi);
 
   appLog.childLog(bundleOpenapi).info("start bundle: ", inputFile.absolutePath);
@@ -41,15 +34,12 @@ export async function bundleParseOpenapi(
     createSpecProcessor({
       mergeAllOf: params?.mergeAllOf,
       ensureDiscriminatorValues: params?.ensureDiscriminatorValues,
+      xOmit: params?.xOmit,
     });
   const bundleResults = await parseOpenapi(inputFile.absolutePath);
-  const parsed: OpenApiBundled = _.isNil(postProcessor)
-    ? bundleResults.bundle.parsed
-    : postProcessor(bundleResults.bundle.parsed);
+  const parsed: OpenApiBundled = _.isNil(postProcessor) ? bundleResults.bundle.parsed : postProcessor(bundleResults.bundle.parsed);
   // todo: CatBase in generic does not get cleaned up...
-  const cleanedParsed: OpenApiBundled = _.isNil(postProcessor)
-    ? parsed
-    : (await doBundle(bundleResults.bundle.source, parsed)).bundle.parsed;
+  const cleanedParsed: OpenApiBundled = _.isNil(postProcessor) ? parsed : (await doBundle(bundleResults.bundle.source, parsed)).bundle.parsed;
 
   // todo: investigate - for pets-simple it will duplicate schemas...
   if (_.isDefined(parsed.components?.schemas?.["schemas"])) {